Move Join The Mission section to its own component then import to welcome and join views
